requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
passlib[bcrypt]
python-jose[cryptography]
websockets
redis
pillow
httpx
